% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exonerate_now.R
\name{ExonerateNow}
\alias{ExonerateNow}
\title{Function to generate BASH scripts that can be used to perform EXONERATE}
\usage{
ExonerateNow()
}
\description{
Function to generate BASH scripts that can be used to perform EXONERATE
}
