% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exonerate_now.R
\name{GetDataForExonerate}
\alias{GetDataForExonerate}
\title{Function that is a part of the MyWitness automation class used to generate all the necessary fasta files}
\usage{
GetDataForExonerate(WitnessObject)
}
\arguments{
\item{WitnessObject:}{Object of class MyWitness}
}
\value{
: Nothing but generates fasta files that are necessary for exonerate
}
\description{
Function that is a part of the MyWitness automation class used to generate all the necessary fasta files
}
