% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gffreader.r
\name{gff2fastaprinter}
\alias{gff2fastaprinter}
\title{Function to print fasta file of the newly acquired CDS co-ordinates}
\usage{
gff2fastaprinter(DeepObject, edited_gff_object = NULL)
}
\arguments{
\item{DeepObject:}{An object of class MyDeep.}

\item{edited_gff_onject:}{Currently set to NULL, but in case you have an edited GFF data, it can be directly used instead.}
}
\description{
Function to print fasta file of the newly acquired CDS co-ordinates
}
