% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNASeq.R
\name{MyBat}
\alias{MyBat}
\title{Constructor and automated function to perform pairwise comparison of RNASeq data.
æparam samples: A character vector with all the kallisto directories of interest in order}
\usage{
MyBat(samples, pairwise_compare = TRUE, kal_dir, conditions, species)
}
\arguments{
\item{kal_dir:}{path to the base directory where all the kallisto directories are present}

\item{conditions:}{A vector that defines the condition of each sample}

\item{species:}{A vector defining he species from which the samples are derived in a one-to-one manner}
}
\description{
Constructor and automated function to perform pairwise comparison of RNASeq data.
æparam samples: A character vector with all the kallisto directories of interest in order
}
