% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gffreader.r
\name{MyDeep}
\alias{MyDeep}
\title{Constructor and automated function of class MyDeep to read the gff output after exonerate and extract CDS data}
\usage{
MyDeep(file, species, style, genomefile)
}
\arguments{
\item{file:}{path to gff file.}

\item{species:}{species that was annotated.}

\item{style:}{Already set to exonerate, can be modified to use for standard gff files.}

\item{genomefile:}{Path to the genomefile from which we can perform the extraction of new CDS data.}
}
\value{
: A list with the gff data, The exon data and finally the fasta data of all the genes(longest isoforms).
}
\description{
Constructor and automated function of class MyDeep to read the gff output after exonerate and extract CDS data
}
