% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exonerate_now.R
\name{MyWitness}
\alias{MyWitness}
\title{Constructor and automated function to read in genome, query and cds files and generate prerequisite files for the step of annotating a genome using Exonerate.}
\usage{
MyWitness(
  blasttype,
  pathtogenome,
  pathtoqueryfile,
  species,
  pathtocdsfile = NULL
)
}
\arguments{
\item{blasttype:}{the type of BLAST you want to perform. All kinds of BLASTing methods are available now}

\item{pathtogenome:}{Path to the genome file(Note: It must be a fasta file)}

\item{pathtoqueryfile:}{Path to the query file for BLAST. It could be different from the CDS file and in case both are same, do not provide any value to pathtocdsfile}

\item{species:}{The species in which we are trying to annotate}

\item{pathtocdsfile:}{Path to CDS file for exonerate in case it is different from the query file for BLAST, initially set to NULL}
}
\value{
: MyWitness object with all the input parameters and generates output BLAST files in CSV format and also generates BASH scripts that can be modified and used for the EXONERATE step
}
\description{
Constructor and automated function to read in genome, query and cds files and generate prerequisite files for the step of annotating a genome using Exonerate.
}
