% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gffreader.r
\name{read_gff}
\alias{read_gff}
\title{Function to read gff file}
\usage{
read_gff(DeepObject, pathtofile = NULL)
}
\arguments{
\item{DeepObject:}{An object of class MyDeep}

\item{pathto}{file: Can also be used separately from the automation and in such a case this parameter can be used to provide input gff file. Initially set to NULL.}
}
\value{
: the data in the gff file as a dataframe
}
\description{
Function to read gff file
}
