% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exonerate_now.R
\name{RemoveFiles}
\alias{RemoveFiles}
\title{Function to remove the fasta and csv files after the completion of Exonerate}
\usage{
RemoveFiles()
}
\description{
Function to remove the fasta and csv files after the completion of Exonerate
}
