% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exonerate_now.R
\name{runBlast}
\alias{runBlast}
\title{Function that is a part of the MyWitness automation class used to perform BLAST}
\usage{
runBlast(WitnessObject, split = 2000, addtitional_args = NULL)
}
\arguments{
\item{WitnessObject:}{Object of class MyWitness}

\item{split:}{A numeric set to split to the csv files for parallelizing the exonerate step}

\item{additional_args:}{Currently set to NULL but user can use it to provide additional arguments for BLAST}
}
\value{
:
}
\description{
Function that is a part of the MyWitness automation class used to perform BLAST
}
